restructure into prefix and registration agency
