+++ libfulltext/doi/crossref/aps.py
@@ -1,10 +1,10 @@
-"""SpringerNature publisher module"""
+"""American Physical Society publisher module"""
 
 import requests
-from .stream import save_to_file
+from ...stream import save_to_file
 
-def get_springer_fulltext(metadata):
-    """Retrieve SpringerNature fulltext
+def get_aps_fulltext(metadata):
+    """Retrieve APS fulltext
 
     Args:
         metadata: meta data dictionary about the DOI. Needs `metadata['message']['DOI']`
@@ -13,7 +13,8 @@ def get_springer_fulltext(metadata):
     doi = metadata['message']['DOI']
 
     response = requests.get(
-        'https://link.springer.com/content/pdf/{0}.pdf'.format(doi),
+        'http://harvest.aps.org/v2/journals/articles/{0}'.format(doi),
+        headers={"Accept": "application/pdf"},
         stream=True
         )
     save_to_file(response)
