+++ libfulltext/doi/crossref/springer.py
@@ -1,10 +1,10 @@
-"""American Physical Society publisher module"""
+"""SpringerNature publisher module"""
 
 import requests
-from .stream import save_to_file
+from ...stream import save_to_file
 
-def get_aps_fulltext(metadata):
-    """Retrieve APS fulltext
+def get_springer_fulltext(metadata):
+    """Retrieve SpringerNature fulltext
 
     Args:
         metadata: meta data dictionary about the DOI. Needs `metadata['message']['DOI']`
@@ -13,8 +13,7 @@ def get_aps_fulltext(metadata):
     doi = metadata['message']['DOI']
 
     response = requests.get(
-        'http://harvest.aps.org/v2/journals/articles/{0}'.format(doi),
-        headers={"Accept": "application/pdf"},
+        'https://link.springer.com/content/pdf/{0}.pdf'.format(doi),
         stream=True
         )
     save_to_file(response)
