+++ libfulltext/doi/datacite/__init__.py
@@ -1,0 +1,5 @@
+"""DataCite DOI handler module"""
+
+def get_datacite_fulltext():
+    """Get fulltext for a DataCite DOI"""
+    raise NotImplementedError('DataCite fulltext retrieval is not yet implemented.')
